float score

+++ examples/bert_predict.py
@@ -326,7 +326,7 @@ def main():
                 output_json["score"] = []
                 for candidate in all_candidates[unique_id]:
                     can_ids=tokenizer.convert_tokens_to_ids(tokenizer.tokenize(candidate))
-                    score=int(scores[can_ids].mean())
+                    score=float(scores[can_ids].mean())
                     output_json["score"].append((candidate,score))
                 writer.write(json.dumps(output_json) + "\n")
